- Inter-address space copy service (XCOPY) - Initial resident shell (rshell), right now just loads and executes   from the file name specified at TASK_BASE - LOADPRC service: create a new process and load an a.out image using rshell

+++ lib/sys/lib_readwrite.c
@@ -2,14 +2,14 @@
 #include "muxxdef.h"
 #include "externals.h"
 
-int write(int fd, int size, char *buffer) {
+int write(int fd, WORD size, char *buffer) {
   PIOTE iote;
 
   iote = curtcb->taskTUCB->iote[fd];
   return _write(iote, size, buffer);
 }
 
-int read(int fd, int size, char *buffer) {
+int read(int fd, WORD size, char *buffer) {
   PIOTE iote;
 
   iote = curtcb->taskTUCB->iote[fd];
